+++ www/go/modules/community/tasks/views/extjs3/TaskGrid.js
@@ -37,6 +37,7 @@ go.modules.community.tasks.TaskGrid = Ext.extend(go.grid.GridPanel, {
 				},
 				'estimatedDuration',
 				'timeBooked',
+				'location',
 				'permissionLevel'
 			],
 			entityStore: this.support ? "SupportTicket" : "Task",
@@ -284,7 +285,16 @@ go.modules.community.tasks.TaskGrid = Ext.extend(go.grid.GridPanel, {
 						return '';
 					},
 					groupable: false
-				}
+				},{
+				id:"location",
+				header: t("Location", "tasks", 'community' ),
+				dataIndex: 'location',
+				align: "right",
+				hidden: true,
+				width: dp(100),
+				groupable: false,
+				sortable: false
+			}
 			];
 
 		if(this.forProject) {